Move consumer factory interface outside connector
