+++ build.gradle
@@ -218,6 +218,7 @@ project(':datastream-kafka-factory-impl') {
 project(':datastream-kafka-connector') {
   dependencies {
     compile project(':datastream-server-api')
+    compile project(':datastream-kafka')
     compile "com.linkedin.kafka.clients:li-apache-kafka-clients:$LIKafkaVersion"
     compile "org.apache.httpcomponents:httpclient:$apacheHttpClientVersion"
     compile "commons-validator:commons-validator:$commonsValidatorVersion"
