+++ datastream-kafka-connector/src/test/java/com/linkedin/datastream/connectors/kafka/TestKafkaConsumerFactoryImpl.java
@@ -12,6 +12,9 @@ import org.apache.kafka.common.serialization.ByteArrayDeserializer;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
+import com.linkedin.datastream.kafka.factory.KafkaConsumerFactoryImpl;
+
+
 /**
  * Tests for {@link KafkaConsumerFactoryImpl}
  */
